Move strtobool into utils
