+++ multiqc/modules/fgbio/ErrorRateByReadPosition.py
@@ -1,8 +1,8 @@
 """ MultiQC submodule to parse output from fgbio ErrorRateByReadPosition """
 
 
-from multiqc.modules.picard.picard import strtobool
 from multiqc.plots import linegraph
+from multiqc.utils.util_functions import strtobool
 
 
 def parse_reports(self):
