+++ multiqc/modules/picard/picard.py
@@ -157,25 +157,3 @@ class MultiqcModule(BaseMultiqcModule):
         except ValueError:
             pass
         return val
-
-
-def strtobool(val):
-    """
-    Replaces deprecated https://docs.python.org/3.9/distutils/apiref.html#distutils.util.strtobool
-    The deprecation recommendation is to re-implement the function https://peps.python.org/pep-0632/
-
-    ------------------------------------------------------------
-
-    Convert a string representation of truth to true (1) or false (0).
-
-    True values are 'y', 'yes', 't', 'true', 'on', and '1'; false values
-    are 'n', 'no', 'f', 'false', 'off', and '0'.  Raises ValueError if
-    'val' is anything else.
-    """
-    val = val.lower()
-    if val in ("y", "yes", "t", "true", "on", "1"):
-        return 1
-    elif val in ("n", "no", "f", "false", "off", "0"):
-        return 0
-    else:
-        raise ValueError("invalid truth value %r" % (val,))
